app/requirements.txt
# fastapi==0.115.0 
# uvicorn==0.30.6 
# sqlalchemy==2.0.35 
# psycopg2-binary==2.9.9 
# pydantic==2.9.2 
# python-jose[cryptography]==3.3.0 
# passlib[bcrypt]==1.7.4 
# python-dotenv==1.0.1
# # pydantic[email]==1.8.0  # Includes email-validator
# email-validator>=2.0.0


fastapi==0.115.0 
uvicorn==0.30.6 
sqlalchemy==2.0.35 
psycopg2-binary==2.9.9 
pydantic==2.9.2 
python-jose[cryptography]==3.3.0 
passlib[bcrypt]==1.7.4 
python-dotenv==1.0.1
email-validator>=2.0.0
python-multipart==0.0.9  # <-- Add this line
pydantic-settings==2.2.1
bcrypt==4.1.3
boto3 == 1.34.119
PyJWT == 2.8.0
requests== 2.32.3
pandas==2.3.0
openpyxl==3.1.5

fpdf==1.7.2